refactor(gateway): preserve transcript usage type ownership
